backend/requirements.txt
Flask
flask-cors
crewai
crewai-tools
python-dotenv
langchain-openai